Temporary fix for SVGs
Groking the code I noticed that Hugo theme is just injecting the svg code in HTML.
I copied their pattern for now since it works. Also appears this is how pol.is is handling their SVGs.

I will read up some more on this benefit over the others. For now, it works.